requirements.txt
numpy
opencv-python
Pillow
argparse
tensorflow
